Move #update_email to /account namespace
The ("Change your email or password") page that uses this endpoint has
already moved to the /account namespace because it's becoming a sub-page
of the new Account page.

I'm undecided on the naming/RESTful design of this resource, so I'm
waiting to take some cues from the upcoming design review in this area.
For the time being I'm keeping it as I found it, but moving it into the
same controller as the page that uses it.
