+++ app/policies/account/email_passwords_policy.rb
@@ -2,4 +2,5 @@ class Account::EmailPasswordsPolicy < BasePolicy
   def show?
     current_user.present?
   end
+  alias_method :update_email?, :show?
 end
